redis缓存user、role逻辑整理
Former-commit-id: 546115e29ba1c66ae992233f5821cb799bb5262c [formerly cadaa31084d64b3360d1352fc9b01c83e0f31884]
Former-commit-id: 0ed22ab64e7cc6c7b5ee2d370b8f3613f9808e18

+++ faber-common/src/main/java/com/faber/core/config/redis/aspect/CacheClearAspect.java
@@ -1,5 +1,7 @@
 package com.faber.core.config.redis.aspect;
 
+import cn.hutool.core.util.ReflectUtil;
+import cn.hutool.core.util.StrUtil;
 import com.faber.core.config.redis.annotation.FaCacheClear;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -48,12 +50,24 @@ public class CacheClearAspect {
 
     private void clearCache(ProceedingJoinPoint invocation, FaCacheClear anno) {
         String pre = anno.pre();
+        String key = anno.key();
 
-        RKeys keys = redisson.getKeys();
-        Iterable<String> keysByPattern = keys.getKeysByPattern(keyPrefix + pre + "*");
-        for (String key : keysByPattern) {
-            RBucket<Object> rBucket = redisson.getBucket(key);
-            rBucket.delete();
+        if (StrUtil.isNotEmpty(key)) {
+            // 这里简单的使用反射获取方法参数值
+            Object[] arguments = invocation.getArgs();
+            Object objValue = ReflectUtil.getFieldValue(arguments[0], key);
+            String fullKey = keyPrefix + pre + StrUtil.toString(objValue);
+            RBucket<Object> rBucket = redisson.getBucket(fullKey);
+            if (rBucket != null) {
+                rBucket.delete();
+            }
+        } else {
+            RKeys keys = redisson.getKeys();
+            Iterable<String> keysByPattern = keys.getKeysByPattern(keyPrefix + pre + "*");
+            for (String k : keysByPattern) {
+                RBucket<Object> rBucket = redisson.getBucket(k);
+                rBucket.delete();
+            }
         }
     }
